move version to root of cmd
